CVSL-1030 move ProbationSearchResult to model as used by returned via API and remove null for probation search result variable
