+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/model/ProbationSearchResult.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation.model
+package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.model
 
 import io.swagger.v3.oas.annotations.media.Schema
