+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/probation/ProbationSearchResponse.kt
@@ -1,5 +1,5 @@
 package uk.gov.justice.digital.hmpps.createandvaryalicenceapi.service.probation
 
 data class ProbationSearchResponse(
-  val content: List<ProbationSearchResult>?
+  val content: List<ProbationSearchResult>
 )
